requirements.txt
pipenv==2022.8.5
discord.py==1.7.3
python-decouple==3.6